pyproject.toml
[tool.poetry]
name = "zaberov-dz-24-1"
version = "0.1.0"
description = ""
authors = ["zaberov.dv"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "4.2.8"
djangorestframework = "^3.14.0"
python-dotenv = "^1.0.1"
ipython = "^8.22.2"
pillow = "^10.2.0"
psycopg2-binary = "^2.9.9"
django-filter = "^24.1"
djangorestframework-simplejwt = "^5.3.1"
coverage = "^7.4.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
